+++ async_include/urls.py
@@ -1,9 +1,12 @@
-from django.urls import path
+try:
+    from django.urls import re_path
+except ImportError:
+    from django.conf.urls import url as re_path
 
 from async_include import views
 
 app_name = 'async_include'
 
 urlpatterns = [
-    path(r'get/?', views.get_template, name="get_template"),
+    re_path(r'^get/?$', views.get_template, name="get_template"),
 ]